fix: remove input require cycle
